Add stopwatch funcitonality
Adds functionality for buttons and time display. Tried to stay as true to the test file as possible but in the 'pauses and resumes' test, the call to .textContent returns undefined because this method does not apply to react-native testing library. I modified the test to grab what I assume was intended (the time xx:xx:xx) and waited 2 seconds before checking the time again to ensure the stopwatch resumes.